+++ src/main/java/tools/devnull/trugger/element/impl/ObjectElement.java
@@ -84,7 +84,7 @@ public final class ObjectElement extends AbstractElement {
       type = setter.getParameterTypes()[0];
       searchForGetter();
     }
-    field = reflect().field(name).in(declaringClass).value();
+    field = reflect().field(name).in(declaringClass).result();
     searchForAnnotatedElement();
   }