+++ src/main/java/tools/devnull/trugger/selector/ConstructorSelector.java
@@ -18,8 +18,8 @@
  */
 package tools.devnull.trugger.selector;
 
-import tools.devnull.trugger.Optional;
 import tools.devnull.trugger.Result;
+import tools.devnull.trugger.SelectionResult;
 import tools.devnull.trugger.reflection.ReflectionException;
 
 import java.lang.reflect.Constructor;
@@ -31,7 +31,7 @@ import java.util.function.Predicate;
  * @author Marcelo Guimarães
  */
 public interface ConstructorSelector extends PredicateSelector<Constructor<?>>,
-    Result<Optional<Constructor<?>>, Object> {
+    Result<SelectionResult<Constructor<?>>, Object> {
 
   ConstructorSelector filter(Predicate<? super Constructor<?>> predicate);
 
@@ -58,6 +58,6 @@ public interface ConstructorSelector extends PredicateSelector<Constructor<?>>,
    *
    * @since 2.1
    */
-  Optional<Constructor<?>> in(Object target) throws ReflectionException;
+  SelectionResult<Constructor<?>> in(Object target) throws ReflectionException;
 
 }