+++ src/test/java/tools/devnull/trugger/reflection/ConstructorSelectorTest.java
@@ -25,13 +25,13 @@ import java.lang.reflect.Constructor;
 import java.util.function.Predicate;
 
 import static org.junit.Assert.assertArrayEquals;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertNull;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.atLeastOnce;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
-import static tools.devnull.trugger.TruggerTest.assertEmpty;
-import static tools.devnull.trugger.TruggerTest.assertNotEmpty;
 import static tools.devnull.trugger.reflection.Reflection.reflect;
 
 /**
@@ -52,13 +52,13 @@ public class ConstructorSelectorTest {
 
 	@Test
 	public void testVisibleSelector() {
-    assertNotEmpty(reflect().visible().constructor().in(TestObject.class));
+    assertNotNull(reflect().visible().constructor().in(TestObject.class).result());
 	}
 
   @Test
   public void testPredicateSelector() {
-    assertNotEmpty(reflect().constructor().filter((c) -> true).withoutParameters().in(Object.class));
-    assertEmpty(reflect().constructor().filter((c) -> false).withoutParameters().in(Object.class));
+    assertNotNull(reflect().constructor().filter((c) -> true).withoutParameters().in(Object.class).result());
+    assertNull(reflect().constructor().filter((c) -> false).withoutParameters().in(Object.class).result());
   }
 
   static class ParameterSelectorTest {
@@ -70,17 +70,17 @@ public class ConstructorSelectorTest {
 
   @Test
   public void testParameterSelector() {
-    assertNotEmpty(reflect().constructor().withParameters(String.class).in(ParameterSelectorTest.class));
-    assertNotEmpty(reflect().constructor().withParameters(int.class, int.class).in(ParameterSelectorTest.class));
-    assertNotEmpty(reflect().constructor().withParameters(double.class, double.class, double.class).in(ParameterSelectorTest.class));
-    assertNotEmpty(reflect().constructor().withParameters(long.class).in(ParameterSelectorTest.class));
+    assertNotNull(reflect().constructor().withParameters(String.class).in(ParameterSelectorTest.class).result());
+    assertNotNull(reflect().constructor().withParameters(int.class, int.class).in(ParameterSelectorTest.class).result());
+    assertNotNull(reflect().constructor().withParameters(double.class, double.class, double.class).in(ParameterSelectorTest.class).result());
+    assertNotNull(reflect().constructor().withParameters(long.class).in(ParameterSelectorTest.class).result());
 
     Constructor<?> constructor = reflect().constructor().withParameters(String.class).in(ParameterSelectorTest.class)
-        .value();
+        .result();
     Class[] parameters = constructor.getParameterTypes();
     assertArrayEquals(new Class[]{String.class}, parameters);
 
-    constructor = reflect().constructor().withParameters(long.class).in(ParameterSelectorTest.class).value();
+    constructor = reflect().constructor().withParameters(long.class).in(ParameterSelectorTest.class).result();
     parameters = constructor.getParameterTypes();
     assertArrayEquals(new Class[]{long.class}, parameters);
   }
@@ -89,11 +89,11 @@ public class ConstructorSelectorTest {
   public void testConstructorPredicate() {
     Predicate<Constructor> predicate = mock(Predicate.class);
 
-    assertEmpty(reflect().constructor().filter(predicate).in(ParameterSelectorTest.class));
+    assertNull(reflect().constructor().filter(predicate).in(ParameterSelectorTest.class).result());
     verify(predicate, atLeastOnce()).test(any(Constructor.class));
 
     when(predicate.test(any(Constructor.class))).thenReturn(true);
-    assertNotEmpty(reflect().constructor().filter(predicate).in(ParameterSelectorTest.class));
+    assertNotNull(reflect().constructor().filter(predicate).in(ParameterSelectorTest.class).result());
     verify(predicate, atLeastOnce()).test(any(Constructor.class));
   }