+++ src/test/java/tools/devnull/trugger/reflection/MethodReflectionTest.java
@@ -23,7 +23,9 @@ import tools.devnull.trugger.Flag;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
-import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.doThrow;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.verify;
 import static tools.devnull.trugger.reflection.MethodPredicates.annotated;
 import static tools.devnull.trugger.reflection.MethodPredicates.annotatedWith;
 import static tools.devnull.trugger.reflection.Reflection.invoke;
@@ -36,7 +38,7 @@ import static tools.devnull.trugger.reflection.Reflection.method;
  */
 public class MethodReflectionTest {
 
-  static interface TestInterface {
+  interface TestInterface {
 
     void doIt();
 
@@ -50,7 +52,7 @@ public class MethodReflectionTest {
   @Test
   public void invokerTest() {
     TestInterface obj = mock(TestInterface.class);
-    invoke(method("doIt")).in(obj).withoutArgs();
+    method("doIt").in(obj).and(invoke());
     verify(obj).doIt();
   }
 
@@ -59,7 +61,7 @@ public class MethodReflectionTest {
     TestInterface obj = mock(TestInterface.class);
     doThrow(new IllegalArgumentException()).when(obj).doIt();
     try {
-      invoke(method("doIt")).in(obj).withoutArgs();
+      method("doIt").in(obj).and(invoke());
       throw new AssertionError();
     } catch (ReflectionException e) {
       assertTrue(IllegalArgumentException.class.equals(e.getCause().getClass()));
@@ -71,29 +73,29 @@ public class MethodReflectionTest {
   @Test
   public void invokerForNoMethodTest() {
     TestInterface obj = mock(TestInterface.class);
-    invoke(method("notDeclared")).in(obj).withoutArgs();
+    method("notDeclared").in(obj).and(invoke());
   }
 
   @Test
   public void predicatesTest() {
     assertFalse(
         annotated().test(
-            method("doIt").in(TestInterface.class).value()
+            method("doIt").in(TestInterface.class).result()
         )
     );
     assertFalse(
         annotatedWith(Flag.class).test(
-            method("doIt").in(TestInterface.class).value()
+            method("doIt").in(TestInterface.class).result()
         )
     );
     assertTrue(
         annotated().test(
-            method("bar").in(TestInterface.class).value()
+            method("bar").in(TestInterface.class).result()
         )
     );
     assertTrue(
         annotatedWith(Flag.class).test(
-            method("bar").in(TestInterface.class).value()
+            method("bar").in(TestInterface.class).result()
         )
     );
   }